Project 4: moved Exception check from getAuth to UserService. Refactored getAuth to take in a string instead of an authToken

+++ server/src/main/java/service/UserService.java
@@ -40,6 +40,9 @@ public class UserService {
     }
 
     public void logout(LogoutRequest request) throws DataAccessException, BadRequestException, UnauthorizedException {
+        if (authDAO.getAuth(request.authToken()) == null){
+            throw new UnauthorizedException();
+        }
         authDAO.deleteAuth(request.authToken());
     }
